revamp project structure
