+++ 1.2.0/scripts/misc/examples.py
