+++ 1.2.0/scripts/train/decorators.py
@@ -5,8 +5,8 @@
 # @Author: Harsha
 # @Email: hvgazula@users.noreply.github.com
 # @Create At: 2024-03-29 09:08:29
-# @Last Modified By: Harsha
-# @Last Modified At: 2024-05-12 10:14:06
+# @Last Modified By: H Gazula
+# @Last Modified At: 2024-05-12 12:01:27
 # @Description:
 #   1. Code to train brainy (unet) on kwyk dataset.
 #   2. binary segmentation is used in this model.
@@ -25,84 +25,21 @@ from icecream import ic
 
 # ruff: noqa: E402
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
-from typing import Dict
 
 import nobrainer
 import tensorflow as tf
-from nobrainer.dataset import Dataset
 from nobrainer.models import unet
 from nobrainer.models.bayesian_meshnet import variational_meshnet
 from nobrainer.processing.segmentation import Segmentation
 from nobrainer.volume import standardize
 
-import label_mapping
-from callbacks import get_callbacks
-from utils import get_git_revision_short_hash, main_timer
+from utils.data_utils import load_custom_tfrec
+from utils.callbacks import get_callbacks
+from utils.py_utils import get_git_revision_short_hash, map_nested_dicts
+from utils.tf_utils import init_device
+from utils.data_utils import load_custom_tfrec
 
 ic.enable()
-
-import collections.abc
-
-
-# https://stackoverflow.com/questions/32935232/python-apply-function-to-values-in-nested-dictionary
-def map_nested_dicts(ob, func):
-    if isinstance(ob, collections.abc.Mapping):
-        return {k: map_nested_dicts(v, func) for k, v in ob.items()}
-    else:
-        return func(ob)
-
-
-def print_config_file(filename):
-    try:
-        with open(filename, "r") as file:
-            config_data = file.read()
-            print(config_data)
-    except FileNotFoundError:
-        print(f"Config file '{filename}' not found.")
-
-
-@main_timer
-def load_custom_tfrec(
-    config: Dict = None,
-    target: str = "train",
-):
-    if target not in ["train", "eval", "test"]:
-        raise ValueError(f"Invalid target: {target}")
-
-    n_classes = config["n_classes"]
-
-    label_map = label_mapping.get_label_mapping(n_classes)
-
-    file_pattern = f"/om2/user/hgazula/kwyk_records/kwyk_full/*{target}*"
-    volumes = {"train": 10331, "eval": 1148}
-
-    if config["debug"]:
-        file_pattern = f"/om2/user/hgazula/nobrainer-data/tfrecords/*{target}*"
-        volumes = {"train": 9, "eval": 1}
-
-    dataset = Dataset.from_tfrecords(
-        file_pattern=file_pattern,
-        volume_shape=config["volume_shape"],
-        block_shape=config["block_shape"],
-        n_classes=n_classes,
-        label_mapping=label_map,
-        n_volumes=volumes[target],
-    )
-
-    return dataset
-
-
-def init_device(flag: bool = False):
-    gpus = tf.config.list_physical_devices("GPU")
-    gpu_names = [item.name for item in gpus]
-    for gpu in gpus:
-        tf.config.experimental.set_memory_growth(gpu, True)
-    NUM_GPUS = len(gpus)
-
-    if flag and not NUM_GPUS:
-        sys.exit("GPU not found")
-
-    return NUM_GPUS, gpu_names
 
 
 def argument_parser():
@@ -135,7 +72,6 @@ def train(func):
         # basic config
         basic_config = config["basic"]
         output_dir = basic_config["model_name"]
-        n_classes = basic_config["n_classes"]
         normalize = basic_config["normalize"]
 
         # training config
