+++ 1.2.0/tests/test.sh
@@ -1,16 +1,16 @@
 #!/bin/bash -l
-python brainy_train.py tests/configs/brainy_01a.yml
-python brainy_train.py tests/configs/brainy_02a.yml
-python brainy_train.py tests/configs/brainy_03a.yml
 
-python kwyk_train.py tests/configs/kwyk_01a.yml
-python kwyk_train.py tests/configs/kwyk_02a.yml
-python kwyk_train.py tests/configs/kwyk_03a.yml
+# test training from scratch (binary and multi-class)
+python scripts/train/brainy_train.py tests/configs/brainy_01a.yml
+python scripts/train/brainy_train.py tests/configs/brainy_02a.yml
 
-python brainy_train.py tests/configs/brainy_01b.yml
-python brainy_train.py tests/configs/brainy_02b.yml
-python brainy_train.py tests/configs/brainy_03b.yml
+python scripts/train/kwyk_train.py tests/configs/kwyk_01a.yml
+python scripts/train/kwyk_train.py tests/configs/kwyk_02a.yml
 
-python kwyk_train.py tests/configs/kwyk_01b.yml
-python kwyk_train.py tests/configs/kwyk_02b.yml
-python kwyk_train.py tests/configs/kwyk_03b.yml
+# test resumption (from checkpoint using decorators)
+python scripts/train/brainy_train.py tests/configs/brainy_01b.yml
+python scripts/train/kwyk_train.py tests/configs/kwyk_01b.yml
+
+# test decorators
+python scripts/train/decorators.py tests/configs/brainy_01a.yml
+python scripts/train/decorators.py tests/configs/kwyk_01b.yml
