+++ 1.2.0/utils/py_utils.py
@@ -1,0 +1,61 @@
+import collections.abc
+import csv
+import subprocess
+from datetime import datetime
+
+__all__ = [
+    "main_timer",
+    "_read_csv",
+    "get_git_revision_hash",
+    "get_git_revision_short_hash",
+    "plot_tensor_slices",
+]
+
+
+# https://stackoverflow.com/questions/32935232/python-apply-function-to-values-in-nested-dictionary
+def map_nested_dicts(ob, func):
+    if isinstance(ob, collections.abc.Mapping):
+        return {k: map_nested_dicts(v, func) for k, v in ob.items()}
+    else:
+        return func(ob)
+
+
+def main_timer(func):
+    """Decorator to time any function"""
+
+    def function_wrapper(*args, **kwargs):
+        start_time = datetime.now()
+        # print(f'Start Time: {start_time.strftime("%A %m/%d/%Y %H:%M:%S")}')
+        result = func(*args, **kwargs)
+        end_time = datetime.now()
+        # print(f'End Time: {end_time.strftime("%A %m/%d/%Y %H:%M:%S")}')
+        print(
+            f"Function: {func.__name__} Total runtime: {end_time - start_time} (HH:MM:SS)"
+        )
+        return result
+
+    return function_wrapper
+
+
+# https://stackoverflow.com/questions/14989858/get-the-current-git-hash-in-a-python-script
+def get_git_revision_hash() -> str:
+    return subprocess.check_output(["git", "rev-parse", "HEAD"]).decode("ascii").strip()
+
+
+def get_git_revision_short_hash() -> str:
+    return (
+        subprocess.check_output(["git", "rev-parse", "--short", "HEAD"])
+        .decode("ascii")
+        .strip()
+    )
+
+
+def _read_csv(filepath, skip_header=True, delimiter=","):
+    """Return list of tuples from a CSV, where each tuple contains the items
+    in a row.
+    """
+    with open(filepath, newline="") as csvfile:
+        reader = csv.reader(csvfile, delimiter=delimiter)
+        if skip_header:
+            next(reader)
+        return [tuple(row) for row in reader]
